requirements.txt
plexapi>=4.15.0
requests>=2.31.0
flask>=3.0.0
python-dotenv>=1.0.0
rich>=13.0.0
humanize>=4.9.0
apscheduler>=3.10.0
